# Add benchmark for json.compact high cpu usage in watch