#### What type of PR is this?

/kind feature

#### What this PR does / why we need it:

#### Which issue(s) this PR fixes:
Adds repro benchmark for https://github.com/kubernetes/kubernetes/issues/110146

It also does tiny refactor of existing benchmark, with no performance diff:
```
name               old time/op  new time/op  delta
WatchHTTP-48       31.5µs ± 0%  33.3µs ± 0%   ~     (p=1.000 n=1+1)
WatchHTTP_UTF8-48  31.9µs ± 0%  32.5µs ± 0%   ~     (p=1.000 n=1+1)
WatchWebsocket-48  32.9µs ± 0%  32.2µs ± 0%   ~     (p=1.000 n=1+1)
WatchProtobuf-48   18.6µs ± 0%  18.9µs ± 0%   ~     (p=1.000 n=1+1)
```

BenchmarkWatchCachingObjectJSON properly reproduces https://github.com/kubernetes/kubernetes/issues/110146:
![image](https://user-images.githubusercontent.com/2559168/197780718-890fd664-3a2a-45fc-b821-434bd0a7ac6a.png)

Such branch isn't visible for BenchmarkWatchCachingObjectProtobuf:
![image](https://user-images.githubusercontent.com/2559168/197780817-7d7d252d-2d30-4edc-86a9-a1a8c53cf6ba.png)



#### Special notes for your reviewer:

It makes CachingObject public so that we can reuse it for performance testing. Alternative approach would be to reimplement some simpler mock, if prefered.

#### Does this PR introduce a user-facing change?
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required:
Enter your extended release note in the block below. If the PR requires additional action from users switching to the new release, include the string "action required".

For more information on release notes see: https://git.k8s.io/community/contributors/guide/release-notes.md
-->
```release-note
NONE
```

#### Additional documentation e.g., KEPs (Kubernetes Enhancement Proposals), usage docs, etc.:

<!--
This section can be blank if this pull request does not require a release note.

When adding links which point to resources within git repositories, like
KEPs or supporting documentation, please reference a specific commit and avoid
linking directly to the master branch. This ensures that links reference a
specific point in time, rather than a document that may change over time.

See here for guidance on getting permanent links to files: https://help.github.com/en/articles/getting-permanent-links-to-files

Please use the following format for linking documentation:
- [KEP]: <link>
- [Usage]: <link>
- [Other doc]: <link>
-->
```docs

```

/assign @wojtek-t (for comments on making CachingObject public)

Wojtek, is it OK to make this public or we prefer to implement fakeCachingObject in the watch_test.go?